newsletter-bot/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
python-dotenv==1.0.1
anthropic==0.28.0
httpx==0.27.0
feedparser==6.0.11
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
jinja2==3.1.4
aiosmtplib==3.0.1
email-validator==2.1.1
pillow==10.3.0
pydantic-settings==2.2.1
pandas==2.2.2
aiofiles==23.2.1
